3.x save preserve publish updates

+++ lib/hyacinth/adapters/storage_adapter/disk.rb
@@ -7,11 +7,8 @@ module Hyacinth
           super(adapter_config)
 
           REQUIRED_CONFIG_OPTS.each do |required_opt|
-            if adapter_config[required_opt].present?
-              self.instance_variable_set("@#{required_opt}", adapter_config[required_opt])
-            else
-              raise Hyacinth::Exceptions::MissingRequiredOpt, "Missing required opt: #{required_opt}"
-            end
+            raise Hyacinth::Exceptions::MissingRequiredOpt, "Missing required opt: #{required_opt}" unless adapter_config[required_opt].present?
+            self.instance_variable_set("@#{required_opt}", adapter_config[required_opt])
           end
         end
 
@@ -35,7 +32,7 @@ module Hyacinth
         end
 
         def delete(location_uri)
-          File.delete(location_uri_to_file_path(location_uri))
+          File.delete(location_uri_to_file_path(location_uri)) if File.exist?(location_uri_to_file_path(location_uri))
         end
 
         def read_impl(location_uri)
